fix: orderStatusId 타입 수정

+++ src/main/java/com/nhnacademy/taskapi/order/repository/OrderStatusRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface OrderStatusRepository extends JpaRepository<OrderStatus, Long> {
+public interface OrderStatusRepository extends JpaRepository<OrderStatus, Integer> {
     Optional<OrderStatus> findByStatusName(String statusName);
